FocoTotal/requirements.txt
fastapi
uvicorn[standard]
databases[sqlite]
aiosqlite
bcrypt
python-jose[cryptography]
python-multipart
aiofiles
